// app/shared/models/pedido/pedido-item-servico-dto.ts
export class PedidoItemServicoDto {
  idPedidoItem: number;
  codigo: string;
  idPedido: number;
  idUsuario: number;
  idTenant: number;
  idContratoCatalogoItem: number;
  contratoCatalogoItem: ContratoCatalogoItem;
  quantidade: number;
  dataEntrega: string;
  dataEntregaPrevista: string;
  entregaProgramada: boolean;
  entregaProgramadaUltimaDataDto: EntregaProgramadaUltimaData;
  datasDasEntregasProgramadas: Array<EntregaProgramada>;
  minDataEntrega: string;
  dataInclusao: string;
  valorTotal: number;
  situacao: SituacaoPedidoItem;
  idPedidoItemPai: number;
  observacao: string;
  valor: number;
  valorLiquido: number;
  valorBruto: number;
  moeda: Moeda;
  idProduto: number;
  produto: Produto;
  idFornecedor: number;
  fornecedor: PessoaJuridica;
  garantia: Garantia;
  frete: TipoFrete;

  idCentroCusto: number;
  centroCusto: CentroCusto;
  idContaContabil: number;
  contaContabil: ContaContabil;
  idAlcada: number;
  idEnderecoEntrega: number;
  enderecoEntrega: Endereco;
  idSubItemSolicitacaoCompra: number;
  usuario: Usuario;

  solicitacaoItem: SolicitacaoItem;
  idIva: number;
  iva: Iva;

  idItemSolicitacaoCompra: number;
  itemSolicitacaoCompra: ItemSolicitacaoCompra;
  idRequisicaoItem: number;
  idCotacaoItem: number;
  idCotacaoRodadaProposta: number;
  loteMinimo: number;
  idUsuarioSolicitante: number;
  descricao: string;
  valorFrete: number;
  prazoEntrega: number;
  faturamentoMinimo: number;
  quantidadeDisponivel: number;
  idComprador: number;
  selected: boolean;
  relateds: Array<PedidoItemServicoDto>;

  //#region NcmImpostos

  ncm: string;
  pisAliquota: number;
  cofinsAliquota: number;
  csllAliquota: number;
  issAliquota: number;
  irAliquota: number;
  inssAliquota: number;

  //#endregion

  constructor(
    idPedidoItem: number,
    codigo: string,
    idPedido: number,
    idUsuario: number,
    idTenant: number,
    idContratoCatalogoItem: number,
    contratoCatalogoItem: ContratoCatalogoItem,
    quantidade: number,
    dataEntrega: string,
    dataInclusao: string,
    situacao: SituacaoPedidoItem,
    idPedidoItemPai: number,
    valor: number,
    valorLiquido: number,
    valorBruto: number,
    moeda: Moeda,
    idProduto: number,
    produto: Produto,
    idFornecedor: number,
    garantia: number,
    frete: TipoFrete,
    idIva: number,
    idItemSolicitacaoCompra: number,
    idSubItemSolicitacaoCompra: number,
    idRequisicaoItem: number,
    idCotacaoItem: number,
    idCotacaoRodadaProposta: number,
    loteMinimo: number,
    prazoEntrega?: number,
    faturamentoMinimo?: number,
    quantidadeDisponivel?: number,
    idComprador?: number,
    idAlcada?: number,
  ) {
    this.idPedidoItem = idPedidoItem;
    this.codigo = codigo;
    this.idPedido = idPedido;
    this.idUsuario = idUsuario;
    this.idTenant = idTenant;
    this.idContratoCatalogoItem = idContratoCatalogoItem;
    this.contratoCatalogoItem = contratoCatalogoItem;
    this.quantidade = quantidade;
    this.dataEntrega = dataEntrega;
    this.dataInclusao = dataInclusao;
    this.situacao = situacao;
    this.idPedidoItemPai = idPedidoItemPai;
    this.valor = valor;
    this.valorLiquido = valorLiquido;
    this.valorBruto = valorBruto;
    this.moeda = moeda;
    this.idProduto = idProduto;
    this.produto = produto;
    this.idFornecedor = idFornecedor;
    this.garantia = garantia;
    this.frete = frete;
    this.idIva = idIva;
    this.idItemSolicitacaoCompra = idItemSolicitacaoCompra;
    this.idSubItemSolicitacaoCompra = idSubItemSolicitacaoCompra;
    this.idRequisicaoItem = idRequisicaoItem;
    this.idCotacaoItem = idCotacaoItem;
    this.idCotacaoRodadaProposta = idCotacaoRodadaProposta;
    this.loteMinimo = loteMinimo;
    this.prazoEntrega = prazoEntrega;
    this.faturamentoMinimo = faturamentoMinimo;
    this.quantidadeDisponivel = quantidadeDisponivel;
    this.idComprador = idComprador;
    this.idAlcada = idAlcada;
  }
}